Insert #sequenced_samples header in PDX MAF
Generate a `#sequenced_samples` header for the dest study's MAF which contains all the samples in the dest study's clinical sample file.

Since all of the samples coming from CMO/BIC/PDX are known to be sequenced, it is safe to assume that the samples in the merged/subsetted clinical sample file were sequenced and belong in the `#sequenced_samples` header in the MAF.

Added test to ensure seq sample header matches expected value.

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>