Remove unused Data Writers for 'New' samples
The CVR fetcher is currently writing data for only the new samples fetched into a tmp directory. This was initially developed with the intention of having GDD process these data files but we never moved forward with this process. GDD was implemented another way instead.

It is my recommendation to remove the steps and classes handling the writing of only new sample data to avoid confusion in the future.

Signed-off-by: Angelica Ochoa <15623749+ao508@users.noreply.github.com>